performance-analysis/requirements.txt
PyYAML==6.0
pandas==1.4.2
matplotlib==3.5.2
pymongo==4.1.1
mysql-connector-python==8.0.29